+++ src/embeddings.jl
@@ -134,14 +134,14 @@ Systems and Turbulence*, Lecture Notes in Mathematics **366**, Springer (1981)
 """
 function reconstruct(s::AbstractVector{T}, γ, τ) where {T}
     if γ == 0
-        return Dataset{1, T}(s)
+        return Dataset(s)
     end
     de::DelayEmbedding{γ} = DelayEmbedding(Val{γ}(), τ)
     return reconstruct(s, de)
 end
 function reconstruct(s::AbstractVector{T}, γ, τ, w) where {T}
     if γ == 0
-        return Dataset{1, T}(s)
+        return Dataset(s)
     end
     de = WeightedDelayEmbedding(Val{γ}(), τ, w)
     return reconstruct(s, de)